+++ tests/test_backends.py
@@ -24,7 +24,7 @@ def my_quantum_function(x, y):  # type: ignore
     qml.RX(y, wires=1)
     qml.CNOT(wires=[0, 1])
     qml.RY(y, wires=1)
-    qml.T(wires=0).inv()
+    qml.adjoint(qml.T(wires=0))
     qml.S(wires=0)
     qml.U1(x, wires=0)
     qml.U2(x, y, wires=0)
@@ -44,7 +44,6 @@ def my_quantum_function(x, y):  # type: ignore
 
 @pytest.mark.parametrize("test_backend", TEST_BACKENDS)
 def test_backends(test_backend: Backend) -> None:
-
     dev = qml.device(
         "pytket.pytketdevice", wires=3, pytket_backend=test_backend, shots=100000
     )